backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-docx
pandas
python-multipart
